DBT/requirements.txt
# DBT and Redshift dependencies
dbt-redshift>=1.7.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

